requirements.txt
google-genai>=1.22.0
lancedb>=0.24.0
pillow>=10.4.0
sentence-transformers>=4.1.0
smolagents[openai]>=1.19.0
tqdm>=4.67.1
transformers>=4.53.0
yt-dlp>=2025.6.25